halscope: Replace deprecated functions for "Select Channel Source" dialog window
Replaces the deprecated functions in the "Select Channel Source" dialog
window.
It now correctly finds the previously selected item.

The search functionality is broken, this can easily be fixed if we
can accept a new entry with a "double click" with the mouse or the "enter key"
The response signal from the dialog window is also changed.