halscope: Make it work with GTK 3
Most deprecated functions is changed, there are four functions left.

There are some issues with this, compared to how it was before.
It uses a lot more resources, I believe that's because it redraws the
window constantly. It shouldn't do that.
The color of the buttons is changed, The change is intended to be temporary,
it looks like we need to use css to format the buttons, css either
inline or in a separate file.
The horizontal trigger line, over the scope, is also not entirely
correct, not sure if it's the window size ore something else.