ClassicLadder: Update the configuration
This also moves the RT target from 'hal/components/CMakeLists.txt'.
Easier to handle the source files when they are in the same place.
Several of the compiler definitions that was only useful to
ClassicLadder has been moved to this CMakeList, less pollution for
all the other targets that's being built.

The source code has also moved to GTK3.