Add function to build Python binaries
There is one TODO in this file, we use 'sed' to update the shebang line,
I'm not sure if this is needed. Looking at the Git history, it's been
around for over 15 years.

With GNU Autoconf the shebang lines will have both the major and minor
Python version:
    #!/usr/bin/python3.11

With the current command in CMake:
    #!/usr/bin/python3