Mark functions as 'constexpr' and 'noexcept'
This commit updates functions related to tool-table.

All functions returned an int, and always the value 0, since it always
is the same value, there isn't any point in returning one.

Annotated the function with 'constexpr' and 'noexcept', downside with
constexpr is that we can't forward declare the them.

Two of the functions is moved to the private scope of the class.